+++ app/notifications/signals.py
@@ -7,7 +7,7 @@ from django.dispatch import receiver
 from ws4redis.redis_store import RedisMessage
 from redis.exceptions import ConnectionError
 
-from helpers_redis import get_redis_publisher
+from helpers.redis import get_redis_publisher
 
 from .models import Notification
